Update README.md
This project presents an advanced chatbot system that can communicate with users and understand the content of their conversations. This system is characterized by its ability to translate between Arabic and English, in addition to the ability to search for information in Wikipedia, know the weather and geographical coordinates, access YouTube, and display the best educational programming sites. The system relies on advanced techniques of machine learning and natural language processing to ensure the quality and efficiency of its interaction with users.